Modify FirstDll

+++ FirstDll/FirstDll/FirstDll.h
@@ -6,4 +6,4 @@
 
 FIRSTDLL_API int fooInDll();
 FIRSTDLL_API int barInDll();
-FIRSTDLL_API const int Pi;
+FIRSTDLL_API const double Pi;
